requirements.txt
networkx==3.3
matplotlib==3.5.2
sqlparse==0.5.1
pytest==8.3.3
sqlglot[rs]
pyyaml